+++ lib/pages/booking/booking_page.dart
@@ -4,7 +4,7 @@ import 'package:capstone_project_mobile/core/model/therapist.dart';
 import 'package:capstone_project_mobile/layouts/my_app_bar.dart';
 import 'package:capstone_project_mobile/core/model/dto/create_appointment.dart';
 import 'package:capstone_project_mobile/pages/booking/booking_list_page.dart';
-import 'package:capstone_project_mobile/pages/layout_page.dart';
+import 'package:capstone_project_mobile/layouts/layout_page.dart';
 import 'package:capstone_project_mobile/core/services/post_service.dart';
 import 'package:capstone_project_mobile/shared/success_screen.dart';
 import 'package:flutter/material.dart';
