+++ lib/pages/therapist_sign_up/upload_certificates_page.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project_mobile/layouts/my_app_bar.dart';
-import 'package:capstone_project_mobile/pages/layout_page.dart';
+import 'package:capstone_project_mobile/layouts/layout_page.dart';
 import 'package:capstone_project_mobile/pages/login/login_email_page.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
